[android][plural] Plurals from android file to master file
Finished the parsing for plurals and transformation of the master file
to support lang:modifier format. This format is simple to read
and allows us to not modified twine too much.
It should be read as:
[name of the string]
lang:modified = string

where lang is the language (e.g en, fr, sp...)
modifier is the plural quantifier such as one, other, zero...