[Android]Fixed an issue with comments
Fixed an issue introduced by the string/plural regex.
Strings and plurals are the 'end of block' markers. Therefore,
a comment followed by a string/plural would be matched as a 'block' (line).

Because of this, the old way of saving the comments for the next string would not work anymore.
We now can try to find a comment for each 'block' of string. Adding the comment works the same after that