[plurals][ios] Parsing stringsdict into masterfile
Added a method that can parse a stringsdict file into the lang:quantity format
This does not support NSStringFormatValueTypeKey yet.
This does not support multi-variable in the plural string yet

Other known limitation: At the moment twine expects the iOS string file
to always be named .*\.strings$
Plurals are stored in .*\.stringsdict$ files.
Twine excepts each string to be stored in one line.