[plurals][apple] Added apple plist formatter
Instead of having everything in the apple formatter, we now have two file formatters.
The idea is to work around the limitations of twine that support only one file format
per platform.
The apple plist will only work on .stringsdict files, which the apple one will work on .strings files